[MLIR] Bubbleup TBD
Currently, the insertion point for bubbling up is after the tensor.pack
this means that the new generic will be created right after the
tensor.pack. This is inconvenient because we are moving the position of
the generic; the idea is to move pack/unpack around, not
linalg.generics. This PR changes the insertion point to preserve the
position of the generic but it has problems.