[MLIR][Bufferization] Early exit on operation without buffer semantics
Buffer deallocation should do an early exit if the operation does not
have buffer semantics or allocate/free effects (i.e., a print). Move the
early exit as a first check in the `verifyOperationPreconditions`
method. This avoids triggering errors on operations that do not have
buffer semantics and, hence, should not be considered in the
deallocation pass.